backend/requirements.txt
spotipy
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-dotenv
bcrypt
passlib
pyjwt
pydantic[email]
pydantic
transformers
torchaudio
requests
python-multipart
python-jose[cryptography]
